CDC/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-tinymce==4.1.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
sqlparse==0.5.1
typing_extensions==4.12.2
